separate helper tests from functions
