+++ package.json
@@ -23,7 +23,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "CI=true react-scripts test --coverage ",
+    "test": "react-scripts test --coverage -u --env=jsdom",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
